Created new classes for all classes in ExcelPivotTableFieldCollection and ExcelPivotTableFieldGroup.
