+++ EPPlus/Table/PivotTable/ExcelPivotTableFieldGroup.cs
@@ -29,8 +29,6 @@
  * Jan Källman		Added		21-MAR-2011
  * Jan Källman		License changed GPL-->LGPL 2011-12-16
  *******************************************************************************/
-using System;
-using System.Globalization;
 using System.Xml;
 
 namespace OfficeOpenXml.Table.PivotTable
@@ -50,142 +48,6 @@ namespace OfficeOpenXml.Table.PivotTable
 			 base(ns, topNode)
 		{
 
-		}
-		#endregion
-	}
-	
-	/// <summary>
-	/// A pivot table field date group.
-	/// </summary>
-	public class ExcelPivotTableFieldDateGroup : ExcelPivotTableFieldGroup
-	{
-		#region Constants
-		private const string GroupByPath = "d:fieldGroup/d:rangePr/@groupBy";
-		#endregion
-
-		#region Properties
-		/// <summary>
-		/// Gets how to group the date field.
-		/// </summary>
-		public eDateGroupBy GroupBy
-		{
-			get
-			{
-				string v = base.GetXmlNodeString(ExcelPivotTableFieldDateGroup.GroupByPath);
-				if (v != "")
-					return (eDateGroupBy)Enum.Parse(typeof(eDateGroupBy), v, true);
-				else
-					throw (new Exception("Invalid date Groupby"));
-			}
-			private set
-			{
-				base.SetXmlNodeString(ExcelPivotTableFieldDateGroup.GroupByPath, value.ToString().ToLower(CultureInfo.InvariantCulture));
-			}
-		}
-
-		/// <summary>
-		/// Gets whether there exist an auto detect start date.
-		/// </summary>
-		public bool AutoStart
-		{
-			get
-			{
-				return base.GetXmlNodeBool("@autoStart", false);
-			}
-		}
-
-		/// <summary>
-		/// Gets whether there exist an auto detect end date.
-		/// </summary>
-		public bool AutoEnd
-		{
-			get
-			{
-				return base.GetXmlNodeBool("@autoStart", false);
-			}
-		}
-		#endregion
-
-		#region Constructors
-		/// <summary>
-		/// Creates an instance of a <see cref="ExcelPivotTableFieldDateGroup"/>.
-		/// </summary>
-		/// <param name="ns">The namespace of the worksheet.</param>
-		/// <param name="topNode">The top node in the xml.</param>
-		internal ExcelPivotTableFieldDateGroup(XmlNamespaceManager ns, XmlNode topNode) :
-			 base(ns, topNode)
-		{
-		}
-		#endregion
-	}
-
-	/// <summary>
-	/// A pivot table field numeric grouping.
-	/// </summary>
-	public class ExcelPivotTableFieldNumericGroup : ExcelPivotTableFieldGroup
-	{
-		#region Constants
-		private const string StartPath = "d:fieldGroup/d:rangePr/@startNum";
-		private const string EndPath = "d:fieldGroup/d:rangePr/@endNum";
-		private const string GroupIntervalPath = "d:fieldGroup/d:rangePr/@groupInterval";
-		#endregion
-
-		#region Properties
-		/// <summary>
-		/// Gets the start value.
-		/// </summary>
-		public double Start
-		{
-			get
-			{
-				return (double)base.GetXmlNodeDoubleNull(ExcelPivotTableFieldNumericGroup.StartPath);
-			}
-			private set
-			{
-				base.SetXmlNodeString(ExcelPivotTableFieldNumericGroup.StartPath, value.ToString(CultureInfo.InvariantCulture));
-			}
-		}
-		
-		/// <summary>
-		/// Gets the end value.
-		/// </summary>
-		public double End
-		{
-			get
-			{
-				return (double)base.GetXmlNodeDoubleNull(ExcelPivotTableFieldNumericGroup.EndPath);
-			}
-			private set
-			{
-				base.SetXmlNodeString(ExcelPivotTableFieldNumericGroup.EndPath, value.ToString(CultureInfo.InvariantCulture));
-			}
-		}
-
-		/// <summary>
-		/// Gets the interval.
-		/// </summary>
-		public double Interval
-		{
-			get
-			{
-				return (double)base.GetXmlNodeDoubleNull(ExcelPivotTableFieldNumericGroup.GroupIntervalPath);
-			}
-			private set
-			{
-				base.SetXmlNodeString(ExcelPivotTableFieldNumericGroup.GroupIntervalPath, value.ToString(CultureInfo.InvariantCulture));
-			}
-		}
-		#endregion
-
-		#region Constructors
-		/// <summary>
-		/// Creates an instance of a <see cref="ExcelPivotTableFieldNumericGroup"/>.
-		/// </summary>
-		/// <param name="ns">The namespace of the worksheet.</param>
-		/// <param name="topNode">The top node in the xml.</param>
-		internal ExcelPivotTableFieldNumericGroup(XmlNamespaceManager ns, XmlNode topNode) :
-			 base(ns, topNode)
-		{
 		}
 		#endregion
 	}
